+++ graylog2-server/src/main/java/org/graylog2/entitygroups/rest/EntityGroupResource.java
@@ -25,9 +25,8 @@ import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.graylog2.audit.AuditEventTypes;
 import org.graylog2.audit.jersey.AuditEvent;
 import org.graylog2.entitygroups.EntityGroupService;
-import org.graylog2.entitygroups.model.EntityCategory;
-import org.graylog2.database.PaginatedList;
 import org.graylog2.entitygroups.model.EntityGroup;
+import org.graylog2.database.PaginatedList;
 import org.graylog2.plugin.rest.PluginRestResource;
 import org.graylog2.rest.models.SortOrder;
 import org.graylog2.shared.rest.resources.RestResource;
